build: use posix_prefix to get installation path
On Debian, Python 3.10 introduced some changes to the scheme
in the system, defaulting to 'posix_local' instead of the 'posix_user'
which was previously used, and is currently used in other Linux
distributions.

For example, these values got changed from (posix_user):
    data = '/usr'
    platlib = '/usr/lib/python3.10/site-packages'
    purelib = '/usr/lib/python3.10/site-packages'
to (posix_local):
    data = '/usr/local'
    platlib = '/usr/lib/python3.10/dist-packages'
    purelib = '/usr/lib/python3.10/dist-packages'

This change forces the usage of the 'posix_user' scheme.

Fixes: PYSIDE-2003
Pick-to: 6.8
Change-Id: Ice0ca9adc37c2652f5ef6ce9e0aec0f2f324206b
Reviewed-by: Shyamnath Premnadh <Shyamnath.Premnadh@qt.io>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>