shiboken6/Documentation: Fully qualify class TOC entries for nested mode
Introduce a data structure for documented classes that has name/full
name and file name so that the class TOC can be sorted by name and
reference the full name. Refactor the fancyToc() function to be able
to handle that.
For the old, flat mode, the unqualified name is used as was before.

Pick-to: 6.8
Change-Id: Ie7528d388faedd5a7bab58394eb6b84db0dc57e4
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>