coin: ci code improvements for build and test
The current patch modifies the structure for the code
in charge of building and testing pyside in the CI.

Instructions within 'coin_test_instructions.py' and
'coin_build_instructions.py' were duplicated, and other slightly
modified, so the code was standarized and placed in
'coin/instructions_utils.py', so the first two files could be heavily
simplified.

A couple of helper functions were removed from build_scripts/utils.py
in order to reduce the code in the general build utilities for the
project.

Pick-to: 6.8
Change-Id: I0cd4bc51edb2e28c5e81a0be8a27be6cb4027bfd
Reviewed-by: Simo Fält <simo.falt@qt.io>