Add QWebEngineDesktopMediaRequest
It was still commented out due to not being present in the 6.7
snapshot used for adaption in
a2486060f78a42739d88a460b47f23b0c75b4a79. Remove private class
QWebEngineMediaSourceModel.

It needs to be a value type since it is used in signal
QWebEnginePage.desktopMediaRequested(const QWebEngineDesktopMediaRequest &),
but has no default constructor. This is normally not supported but works
due to it not being used in any other API.

Task-number: PYSIDE-2497
Fixes: PYSIDE-2913
Pick-to: 6.8
Change-Id: Ifaf6bd11988e802b245b33724b1b0b071f7e098a
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>