doc: Avoid same name for directory and main qml file
Discovered while testing deployment on macOS,
nuitka fails when the directory and file (e.g. Main/Main.qml)
have the same name.

This might be able to be fixed by other arguments or even
upstream in Nuitka, this is a simple change that will encourage
people not to use the same name in the meantime

Pick-to: 6.8
Fixes: PYSIDE-2919
Task-number: PYSIDE-2910
Change-Id: I596d820e4a30996f9cf934a9f93e6436e51ad0d2
Reviewed-by: Shyamnath Premnadh <Shyamnath.Premnadh@qt.io>
Reviewed-by: Christian Tismer <tismer@stackless.com>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>