Delete QQmlApplicationEngine on application exit
Due to the API limitations, we have to ensure that the engine is deleted
before other parts of the application is deleted. Otherwise exposing
objects using setInitialProperties() or setContextProperty() for example
will cause warnings to be printed. It is a good practice to always
delete the engine manually so all the code should be consistent.

Task-number: PYSIDE-1612
Pick-to: 6.8
Change-Id: I01f16359e9d90cefd5957708fe12ce489bd7edc0
Reviewed-by: Jaime Resano <Jaime.RESANO-AISA@qt.io>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>