shiboken6/Documentation: Prominently display QML types
Previously the link to the QML types list (Qt web documentation) was
automatically inserted into the last paragraph of the module
description. This only had an effect when no PySide module description
was present. Otherwise, it had to be manually added to the module
description which was overlooked for some modules.

To fix this, extend the module documentation obtained from the doc
parsers by the URL to page and format it as a separate section
preceding the class list.

Remove the QML types from the PySide module descriptions.

Pick-to: 6.8
Task-number: PYSIDE-1106
Change-Id: I0b107a81b239405c0258ecff924aef87472f13b0
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>