type hints: Pin mypy to version 1.13 to unblock integration
mypy 1.14 produces errors like:

pyside6/PySide6/QtCore.pyi:1251: error: Detected enum
"PySide6.QtCore.QCborTag" in a type stub with zero members. There is a
chance this is due to a recent change in the semantics of enum
membership. If so, use `member = value` to mark an enum member,
instead of `member: type` [misc]

pyside6/PySide6/QtGui.pyi:6866: error: Detected enum
"PySide6.QtGui.QRhi.BeginFrameFlag" in a type stub with zero members.
There is a chance this is due to a recent change in the semantics of
enum membership. If so, use `member = value` to mark an enum member,
instead of `member: type` [misc]

Task-number: PYSIDE-2846
Pick-to: 6.8
Change-Id: I93d3a03942f86472c0373d622122879172b897ff
Reviewed-by: Christian Tismer <tismer@stackless.com>