type hints: Adapt to the new typing rule of disallowing empty enums
Error = """\
error: Detected enum "PySide6.QtCore.QCborTag" in a type stub with
zero members. There is a chance this is due to a recent change in
the semantics of enum membership. If so, use `member = value` to
mark an enum member, instead of `member: type`  [misc]
"""
So the assumption seems to be that we use wrong syntax if we have
no enum at all. This should be reported as a bug.

Fixed by triggering an error ignore. (Others ignore the empty enum
completely, but then we would need to change all referring classes)

Task-number: PYSIDE-2846
Pick-to: 6.8
Change-Id: I46cc60f9e5537bd6c37c0d34ee3b5b985b0cdc37
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>