pyside6-project: 4. Enable building Design Studio projects
pyside6-project tool needs to be more flexible to build Design Studio
projects. The path where a .qrc compilation output is stored has to be
configurable. In addition, a check is performed to see whether the .qrc
file needs to be rebuild by looking at the files that are referenced.
This avoids unnecesary rebuilds.

Task-number: PYSIDE-1612
Pick-to: 6.8
Change-Id: I8aafd5e66e477bbb360f3cf691d43e12c8866ec4
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>