Change binding for QSignalSpy to Py_DECREF wrapped type
This is arguably a bug, but would only be noticed in testing.  It was
preventing proper garbage collection in tests using QSignalSpy.

I can see an argument it should keep the target object alive while
running, but I don't see a good way to DECREF on close, so not holding
a reference seems like the best option.

Change-Id: I730d8b0a6efd3fef63dfd224917c2e5927d144c4
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>