shiboken6: Reuse existing implementations of Python overrides
Add functionality to generate the code handling Python overrides
into separate static functions of the wrappers.

From the cache of reusable Python overrides, generate a per-class
mapping of virtual functions to functions of other classes which can
be reused.  If a mapping exists for the function, include the header
of the source wrapper and call the function. Otherwise, generate the
static helper into the wrapper.

Size improvement (compared to unmodified branch):

QtCore      4145160      4034568    2.67%
QtGui       4996328      4959464    0.74%
QtWidgets   7202520      6444760    10.52%

Comparing to folding:
QtCore      4145160      4046888    2.37%
QtGui       4996328      4979944    0.33%
QtWidgets   7202520      6481624    10.01%

Task-number: PYSIDE-2701
Change-Id: I7d74c97ccf1aa0cc23c6d86628c6f4eded0a289c
Reviewed-by: Shyamnath Premnadh <Shyamnath.Premnadh@qt.io>