binary size: Update the benchmark tool to absolute measurement
After sizebench.py was used successfully on relative changes
(with a disable switch in the code), we now add an absolute
mode. This way, we can stop to support switching.

The starting point is 2025-01-27 right before a substantial
change which is otherwise not measurable.

The new mode is activated with --absolute / -a

It works much faster since it compares against a constant.

Task-number: PYSIDE-2701
Change-Id: I341a23722a304473e92f8ef38c6d16b3b69a8a92
Pick-to: 6.8
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>