type hints: Replace removal of inconsistent overloads by error ignore
When switching to positional-only parameters, a few new errors
popped up when overloads had inconsistent methods and staticmethod.

This was not perfectly solved by removal. Instead, we keep this
inconsistency and disable the MyPy errors. This is not only a misc
error, but an overload-cannot-match error follows immediately.

Hopefully Qt will drop these mixed cases, soon.

Change-Id: If8e9c2cf57f1a44ae7db9985532c191c22d1a64e
Fixes: PYSIDE-3003
Task-number: PYSIDE-2846
Pick-to: 6.8
Reviewed-by: Ece Cinucen <ece.cinucen@qt.io>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>