Add .gitreview file
`git-review` is a command line tool for interacting with Gerrit.
Some IDEs offer support for this tool, either natively or through
plugins. The tool relies on a configuration file, `.gitreview`. Adding
this configuration file to our repositories simplifies initial setup for
contributors that use the tool directly or through their IDE of choice.

The configuration file adds a remote called "gerrit". This is the
default for the tool, and also the name set for
`codereview.qt-project.org` by Qt's `init-repository` script. Thus,
the configuration should work seamlessly alongside other repository
helpers.

Pick-to: 6.8
Task-number: QTBUG-132604
Change-Id: I01d8dc9cdb1fe4cc32f52ad08dd4146a735193a6
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Paul Wicking <paul.wicking@qt.io>