shiboken6: Add a way of removing redundant overloads that differ in one parameter
When retrieving the function groups, check if there overloads that
differ in one parameter and remove them based on rules specified in
the type system. For example, when QByteArray::number(qint64) and
QByteArray::number(int) exist, it does not make sense to generate an
overload for int with parameter checking.

[ChangeLog][shiboken6] It is now possible to specify rules
to remove redundant overloads that differ in one parameter.

Task-number: PYSIDE-2701
Change-Id: I8e1ccbccfe7b5ad76486dc197e8d9c36f1809fbe
Reviewed-by: Shyamnath Premnadh <Shyamnath.Premnadh@qt.io>