pyproject.toml: 2. Add pyproject.toml support for pyside6 tools
This patch adds support for pyproject.toml files to the pyside6-project
tool. A new command argument is added to migrate a .pyproject JSON file
to the new pyproject.toml file:
`pyside6-project migrate-pyproject`

The new features are tested and it is guaranteed that the current
behavior is preserved. A new flag is added to the project creation operations, "--legacy-pyproject", in order to generate a .pyproject file instead of a pyproject.toml file.

Note that the tomlkit library is added to the requirements.txt file.
https://github.com/python-poetry/tomlkit

Task-number: PYSIDE-2714
Change-Id: If33956dea73b79df0a52d4dcda3934c85e57182d
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>