build: reflect licenses in the wheel description
Qt is released under LGPLv3 or GPLv2 or GPLv3 and PySide as well,
but for the simplicity of writing the description in the README
and package description, LGPL/GPL or LGPLv3 has been written,
which has caused confusion when dealing with GPL-only dependencies.

To avoid this problem, we now reflect the licenses that has been
in the repository and each file in the package description (PEP 639).

Change-Id: I1e24c39ac415e635f5c44b57ccac0175e13b3736
Pick-to: 6.8 6.5 6.2
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Ece Cinucen <ece.cinucen@qt.io>
Reviewed-by: Christian Tismer <tismer@stackless.com>