libshiboken: Change wrapper map to be a multimap to allow for colocated objects
In the case of objects that have the same address in C++ (which
happens for example for the first field of a struct or unions),
entries in the wrapper map would overwrite each other, which requires
complicated handling in libshiboken and the generated code. Change the
wrapper map to be a multimap and add functions searching by type and
instance. This also makes addding support for C++ unions (like
QQuickItem::ItemChangeData) easier.

Task-number: PYSIDE-2854
Change-Id: If8fb5ee43d3285fd25e471ce8c9a3e6ae37ae311
Reviewed-by: Shyamnath Premnadh <Shyamnath.Premnadh@qt.io>