type hints: Fix an old typo in typesystem_core_common.xml
Instead of typing.Iterable, typing.List must be used in
the return type of QtCore.findchildren.

As a drive-by, the name "PlaceholderType" was corrected
because "placeholder" is only one word.

The new calculation of Optional replacement was also
fixed to handle return types.

Note: The Optional[PlaholderType] is special, and typing
refuses to use "|" here.

Change-Id: I9f7d89027496f3cae58710595d7bcaafb0e51915
Fixes: PYSIDE-3045
Pick-to: 6.8 6.9
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>