feature: Fix regression when PySide is imported with trace set
The feature switching redirects imports in order to support the
"from __feature__ import" construct. When no feature is involved,
the original function and some post processing is called.

When settrace is used before the Shiboken import, we see infinite
recursion in Python 3.13. An optimization in updatecache now calls
import itself, which then runs the feature machinery recursively.

Instead of only calling the original import, temporarily switch
the whole import redirection back to the original.

Change-Id: Ie3c2e24467ca456b0e731db2696a0b2ab80c6174
Fixes: PYSIDE-3054
Pick-to: 6.9 6.8 6.8.3
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>