pytest: Fix a small glitch in _mod_uses_pyside: revisited
This was a regression concerning feature.py where a problem
in _mod_uses_pyside led to an all Exception case.
This had the bad effect that an infinite recursion in
Python 3.13 was never seen until this was discovered
my Bernhard Dietermayr.

The handling of the feature import has recently been greatly
improved so that recursion is no longer possible.

Unfortunately, the Pytest problem could not be reproduced,
but this catch-all exception must be removed urgently.

Task-number: PYSIDE-2393
Task-number: PYSIDE-3054
Change-Id: I0fb42752dacd882b313ceacf7d591cd9c8ce16a6
Pick-to: 6.8 6.9
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>