Fix crash retrieving a QGraphicsProxyObject from QVariant
Remove the default-superclass specification QGraphicsItem for
QGraphicsObject since it causes the type discovery to directly cast
from QGraphicsObject* to QGraphicsItem*. This crashes
since QGraphicsObject inherits from QObject as first base class.

The intention of the class attribute was to ensure that any
QGraphicsObject class is stored as a QGraphicsItem. To preserve this,
hardcode it in QVariant_resolveMetaType().

Pick-to: 6.9
Fixes: PYSIDE-3069
Task-number: PYSIDE-86
Task-number: PYSIDE-1887
Change-Id: I3704988f105b118b1e4ef8d078b68c01ba89386c
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>