type hints: Improve topological sorting between Sequence and Iterable
Sorting between Sequence and Iterable defaulted to alphabetical.
For an order of generality, the sorting order needs to be defined.

The reason for this was a problem with the 6.10 branch of PySide
which has enhanced NumPy support. Note that this re-ordering of
Sequence and Iterable is for PYI files, only. The NumPy support
code has completely different needs for sorting that is unrelated.

Task-number: PYSIDE-3098
Change-Id: I2325d4c0a7aa6dbac8337a03e466d992ada5cb06
Pick-to: 6.8 6.9
Reviewed-by: Ece Cinucen <ece.cinucen@qt.io>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>