type hints: Bump mypy version and support disable-gil
There has been extra speedup tried using special compilation options.
These do not work with disable-gil because at the moment disable-gil
prevents compilation of mypy.

It is not possible to specify disable-gil in requirements.txt as a
condition. Because version 1.15 is reasonably faster, it is better
to remove the mypy[faster-cache] option.

Change-Id: I4f846d8ce91d6804d0834d9a1ad819278833644c
Task-nunber: PYSIDE-2221
Task-number: PYSIDE-2846
Pick-to: 6.9
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>