Modify headers installation for CMake builds
Previously, the headers are installed as ${CMAKE_INSTALL_PREFIX}/include/cmake_package_name,
where cmake_package_name is the CMake package name, e.g., PySide6. In
the wheels, the headers were included as package_name/include. Since
the level of the include directory is different, this caused issues
when importing the CMake package in a downstream project.

- This change modifies the installation of the headers so that the
  CMake install prefix and the wheels are consistent.
- Additionally
  - this change adds the libpyside6 headers to the wheels.
  - The include headers for libshiboken were shipped with
  shiboken6-generator and not with shiboken6, which has the
  libshiboken binary.

Change-Id: I96e3280799da169836e24551b906274f5b0fc962
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>