Remove hard-coded Q_OS_ defines from the CMake files
The hard-coded ENABLE_ and Q_OS_ do not work for cross-compiling. They
are in fact a left-over from the pre-clang code parser. Clang is
actually able to set the right defines when parsing qsystemdetection.h
if the correct target is set.

pyside6_global.h then no longer needs to be configured by CMake.

Task-number: PYSIDE-3105
Change-Id: I640ab9e254b9ba35d5b7194001f5700911a8d421
Reviewed-by: Shyamnath Premnadh <Shyamnath.Premnadh@qt.io>