Avoid using typing.Self due in versions < 3.11
When typing.Self doesn't exist (< 3.11) we were monkey-patching
the attribute to the typing module, in order to use 'typing.Self'
for our constructors stubs.

This was affecting the behavior of another module, that was performing
a similar check on the typing.Self attribute.

Amends 1ef1fefc26038a80ba81a860cff5024db44dca37

Change-Id: I5bacbb3db6a7ecd7c10fcbf08f702172e5163868
Fixes: PYSIDE-3135
Pick-to: 6.9 6.8
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>