libpyside: Fix corruption connecting several signals to one non-QObject receiver
The deletion of the weak ref in ~TrackingMethodDynamicSlot() did not
reliably stop the notification from being triggered, causing
corruption later on since the address of the dynamic slot was used as
user data.

Use the Python receiver object as user data instead (which is only
used for map look up) to delete the connection to protect against
multiple invocation.

Remove the deletion of the weakref since it is removed in the
notification.

Remove class TrackingMethodDynamicSlot as it does not really have any
functionality any more.

The comment about releasing the weakref in case DynamicSlot outlives
Python is apparently not an issue since otherwise the deletion of the
function in ~MethodDynamicSlot() would have caused issues.

Amends 33bd61d13d8d9e3794b6049891be62f3351313d9.

Pick-to: 6.9 6.8
Fixes: PYSIDE-3148
Task-number: PYSIDE-2810
Change-Id: Idc07d0774afaf99df93185c90e975291a42ffeaf
Reviewed-by: Shyamnath Premnadh <Shyamnath.Premnadh@qt.io>