Introduce multi-phase Python extension module initialization
For the non-Nuitka case, use the new form of of module initialization,
changing the module init() function to call PyModuleDef_Init()
providing an exec()-function as slot to be called for each
subinterpreter,

For the Nuitka case, maintain the old code path and call the exec()
function directly, as  Nuitka does not seem to support it yet.

At the moment, the lazy initialization code is called from the exec()
function. Full subinterpreter support requires further investigation
on how to make lazy initialization work with multiple binterpreters.

[ChangeLog][shiboken6] Multi-phase Python extension module initialization
is now used for the non-deployed case.

Task-number: PYSIDE-3155
Task-number: PYSIDE-2221
Change-Id: Ie55953d47354b8f832d935fe42a686739dbce2ca
Reviewed-by: Shyamnath Premnadh <Shyamnath.Premnadh@qt.io>