Reduce usage of PyTypeObject::tp_name
In the Limited API, PyTypeObject is an opaque struct, for which
libshiboken provides a dummy definition.

PyType_GetFullyQualifiedName() (stable API since 3.13) can be used as
a replacement, but it returns a PyObject.

Add a convenience function PepType_GetFullyQualifiedNameStr() similar
to the existing PepType_GetNameStr() to return a C-string.

Leave the 3.13 code commented out for the moment since it causes a
crash.

This does not cover occurrences of tp_name passed as strings
to Python formatting functions using the %s placeholder since that
can be replaced by the new %N/%T placeholder starting from 3.13.

Task-number: PYSIDE-3171
Change-Id: I4a073872cd0d138b8d8c6aafb08ccb33451812ca
Reviewed-by: Shyamnath Premnadh <Shyamnath.Premnadh@qt.io>