Documentation: Use a collapse directive for the detailed class description
Traditionally, the class pages were modelled after the Qt C++
documentation. There was a brief description, ending with a reference
to the detailed explanation, labelled "more..".

While this actually works in sphinx, too, it starts to emit warnings
about multiple definitions of the 'more' label.

To fix this, introduce a sphinx collapsible directive, expanding to
the detailed description. This however only works unless the detailed
description has headlines (for example, for QObject).

This creates a difference in that the detailed description is always
initially hidden, but it is assumed that this is better for the common
use case of browsing for a certain method.

Pick-to: 6.10 6.9 6.8
Task-number: PYSIDE-1106
Change-Id: I152be16f2a5ee5ea8c2869396a207b554738b219
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>