shiboken6: Add typedef'ed enumerations
Traverse the typedefs of a scope and check whether they point to some
enumeration and whether a type entry for the source exists. Add these
like normal enums. In the converter type check, add a clause for the
alias source/target type, allowing for using the values
interchangeably.

[ChangeLog][shiboken6] shiboken6 can now also generate typedef'ed
enumerations for purposes of renaming/deprecating enumerations.

Pick-to: 6.10
Change-Id: Iec10c53b1167b958647242cedb04f8ff01ad1085
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>