libshiboken/lazy loading: Refactor incarnateType
When instantiating nested subtypes of a type like
"QTimeZone.OffsetData", incarnateHelper() used to look up "QTimeZone"
in the module dictionary to add the subtype to the dictionary. This
caused a recursion with subinterpreters fixed by
fbf59240583d85e4fbc5f66c9de597ece8759ce4.

To fix this in a better way, split the incarnateHelper() apart such that
the enclosing type can be passed as a parameter and pass the main type
from incarnateSubtypes() directly. Also remove subtypes from func map
when incarnating.

Task-number: PYSIDE-3155
Task-number: PYSIDE-2404
Change-Id: Ib578743332ac03dd120724a776c7edca4b381037
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>