Fix handling type with equal names in signal/slot
The decision whether to create a derived meta object class parsed from
the Python type in MetaObjectBuilder was based on comparing the class
name to the base meta object (the Python parsing must not be done
when creating a meta object for a plain wrapped Qt type).

This led to mixups when base class names were identical which is
possible in Python.

To fix this, split apart the code path not requiring Python type
parsing (called from the wrapper code when initializing wrapped Qt
classes) and expand the code for Python derived classes into
initQObjectSubType().

Fixes: PYSIDE-3201
Pick-to: 6.10
Change-Id: Id62e7dc9b8af16154b161cdbf5dd0d460c55f6f1
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>