shiboken6: Reduce Sbk_..TypeF() function usage
A static helper function with postfix TypeF() was generated for each
type. Besides the usage in the generated converter and for overridden
functions, it was used to retrieve the enclosing type for enum
initialization.

The calls from the enum code are removed by adding an enclosing
parameter to the enum function and removing the logic trying to detect
invisible namespaces since the enclosing type is always the right one
even for enums collected from nested invisible namespaces.

The generation of the TypeF() can then be removed for namespace
types.

Task-number: PYSIDE-3155
Change-Id: Ia1c64594e180cf221cd0c416ae67ddab5d99a465
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>