Fix tests using the Wayland platform plugin
Remove the loop ensuring the window is mapped introduced by
918cd72f5eaebf56287a5ab0ac2e149ba1617a57, since on Wayland, the paint
event occurs there, causing a hang. This loop is usually used to
ensure a window is in a deterministic state for playing mouse events,
etc, but for paint events it should not be required.

Skip the native event test for Wayland, too.

Change-Id: I2b44dc098873614e2b8d9bc54444917fd4905e9f
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>