Re-port the player example
Adapt to the current state of C++ example, which adds the audio level
meter and meta data display. The playlist handling was again removed
since it still relies on a copy of the QMediaPlaylist class from Qt 5.
Simplify the mime type handling to default to MP4 always, which should
work when using FFMPEG.

Task-number: PYSIDE-2206
Pick-to: 6.10
Change-Id: I6215c4bcff5a3ca83606e33da6da475caf6494d8
Reviewed-by: Shyamnath Premnadh <Shyamnath.Premnadh@qt.io>