shiboken6: Refactor and fix the synthesizing of comparison operators
Change the way it is implemented to traverse the operator<=>()
function and construct a AbstractMetaFunction from it first, so
that the standard shiboken rejection mechanism works and
the argument is properly converted.

Change AbstractMetaClass::addSynthesizedComparisonOperators()
to take an AbstractMetaArgumentList and a mask of operators
to be created, enabling synthesizing of non-homogeneous
operators and operator!=() from operator==() in a subsequent change.

Implement the rule for synthesizing operator==()/operator!=()
from operator<=>() correctly (only happens for defaulted operator<=>()).

Do not synthesize ordering operators for non-pair type containers.

Task-number: PYSIDE-3245
Change-Id: I83865253c05ea87ebd874f0c619115f132a98477
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>