Fix top-level configuration for IDE
Shiboken6Config.abi3.cmake checks to include Shiboken6Targets.cmake
if the Shiboken6::libshiboken does not exist, but it should always
exist in a top-level build.

Re-add the Shiboken6::libshiboken alias.

It is unclear why it was removed in the first place, so just to be
safe, only create the alias if it does not already exist.

Amends 39bf64f845d6b759ebc726e85949e82dab678d37

Change-Id: I1d5f29cd770fde7288dcd27703304b8fa5d1f2f5
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>